% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R
\name{cor.test.fisherZ}
\alias{cor.test.fisherZ}
\title{cor.test.fisherZ}
\usage{
\method{cor.test}{fisherZ}(r1 = NULL, r2 = NULL, n1 = NULL, n2 = NULL,
  p = TRUE)
}
\arguments{
\item{r1}{First correlation}

\item{r2}{Second correlation}

\item{n1}{First sample size}

\item{n2}{Second sample size}

\item{p}{Compute p-value? (default = TRUE)}
}
\description{
cor.test.fisherZ
}

