% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ML2.S1}
\alias{ML2.S1}
\title{ManyLabs2 - Slate 1}
\format{A data frame with 7263 rows and 152 variables:
\describe{
  \item{.id}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Source.Global}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Source.Primary}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Source.Secondary}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Country}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Language}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Execution}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{SubjectPool}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Setting}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Tablet}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Pencil}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{StudyOrder}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{IDiffOrder}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{StudyOrderN}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{IDiffOrderN}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{huan1.1_Y1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{huan1.1_R0}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{huan2.1_Y1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{huan2.1_R0}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{homewealth}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{kay1.4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{kay1.5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{kay1.6}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{kay2.4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{kay2.5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{kay2.6}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{alt1.3}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{alt1.4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{alt1.5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{alt1.6}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{alt1.7}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{alt1.8}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{alt2.3}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{alt2.4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{alt2.5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{alt2.6}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{alt2.7}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{alt2.8}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.8}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.9}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.10}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.11}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.12}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.13}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.14}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.15}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.16}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{politics}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.2}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.3}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.6}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{grah1.7}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{rott1.1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{rott2.1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{bau1.1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{baudv.4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{bau2.1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{baudv.2}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{baudv.3}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{baudv.5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{baudv.6}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{miya2.4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{miya2.7}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{miya1.4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{miya1.7}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{miya2.5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{miya1.5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{miya2.6}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{miya1.6}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{miya2.8}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{miya1.8}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{inba1.3}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{disg1.11}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{disg1.12}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{disg2.10}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{disg2.12}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{disg2.13}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{inba2.3}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{inba1.4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{inba2.4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{inba1.5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{inba2.5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{crit1.1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{crit2.1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{van.p1.2_1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{van.p1.2_2}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{van.p1.2_3}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{van.p1.2_4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{van.p1.2_5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{van.p1.2_6}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{van.p2.1_1_TEXT}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{van.p2.1_2_TEXT}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{haus1.1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{haus1.1t_1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{haus2.1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{haus2.1t_1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{haus1.2}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{haus2.2}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.3_1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.3_2}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.3_3}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.3_4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.3_5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_2}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_3}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_6}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_7}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_8}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_9}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_10}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_11}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_12}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_13}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_14}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_15}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_16}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_17}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_18}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_19}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and1.4_20}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.3_1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.3_2}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.3_3}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.3_4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.3_5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_1}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_2}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_3}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_4}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_5}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_6}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_7}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_8}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_9}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_10}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_11}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_12}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_13}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_14}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_15}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_16}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_17}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_18}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_19}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{and2.4_20}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{ross.s1.2}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{ross.s1.1_1_TEXT}{ML2 Slate1 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
}}
\source{
\url{https://osf.io/8cd4r/}
}
\usage{
ML2.S1
}
\description{
A dataset containing results from the first slate of the ManyLabs2 project:
}
\details{
Klein, R. A., Vianello, M., Hasselman, F., ..., Nosek, B. et al. (2016, March 15).
Many Labs 2: Investigating Variation in Replicability Across Sample and Setting. Retrieved from osf.io/8cd4r
}
\keyword{datasets}

