% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ML2.S2}
\alias{ML2.S2}
\title{ManyLabs2 - Slate 2.}
\format{A data frame with 7994 rows and 225 variables:
\describe{
  \item{.id}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Source.Global}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Source.Primary}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Source.Secondary}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Country}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Language}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Execution}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{SubjectPool}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Setting}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Tablet}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{Pencil}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{StudyOrder}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{IDiffOrder}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{StudyOrderN}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{IDiffOrderN}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{ross.s2.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{ross.s2.1_1_TEXT}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{geis1.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{geis.dv_1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{geis.dv_2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{geis.dv_3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{geis.dv_4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{geis.dv_5}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{geis2.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{tver1.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{tver2.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{haus3.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{haus3.1t_3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{haus4.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{haus4.1t_3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{haus3.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{haus4.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{rise1.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{rise2.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sex}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.5}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.9}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.10}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.15}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.16}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.21}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.22}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.27}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.28}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.33}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.34}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.38}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.39}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.43}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava1.44}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.5}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.9}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.10}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.15}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.16}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.21}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.22}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.27}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.28}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.33}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.34}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.38}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.39}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.43}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{sava2.44}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.5}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.6}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.7}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.8}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.9}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.10}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.11}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.12}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.13}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.14}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.15}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.16}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.17}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.18}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.19}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore1.20}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.5}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.6}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.7}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.8}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.9}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.10}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.11}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.12}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.13}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.14}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.15}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.16}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.17}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.18}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.19}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nore2.20}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{hsee1.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{hsee2.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gray1.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gray2.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gray1.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gray2.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gray1.4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gray2.4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nCopied.zhon2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zhon.dv.1_1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zhon.dv.1_2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zhon.dv.1_3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zhon.dv.1_4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zhon.dv.1_5}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zhon.dv.1_6}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zhon.dv.1_7}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zhon.dv.1_8}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zhon.dv.1_9}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zhon.dv.1_10}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{nCopied.zhon1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{schw1.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{schw1.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{schw2.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{schw2.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{shaf1.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{shaf2.1}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zav.dv.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zav.include.strict}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zav.condition}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{zav.dv.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{knob1.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{knob2.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{knob1.4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{knob2.4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.5}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.11}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.12}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.14}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.16}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.17}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.18}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.19}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.22}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.6}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.7}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.8}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.9}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.10}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.13}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.15}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.20}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1s.21}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.5}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.11}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.12}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.14}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.16}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.17}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.18}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.19}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.22}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.6}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.7}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.8}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.9}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.10}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.13}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.15}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.20}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2s.21}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.5}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.11}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.12}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.14}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.16}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.17}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.18}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.19}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.22}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.6}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.7}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.8}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.9}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.10}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.13}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.15}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.20}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati1d.21}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.2}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.3}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.5}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.11}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.12}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.14}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.16}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.17}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.18}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.19}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.22}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.4}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.6}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.7}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.8}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.9}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.10}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.13}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.15}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.20}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
  \item{gati2d.21}{ML2 Slate2 variable: See column \code{study.vars} in the \href{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}{MasterKey spreadsheet} for details.}
}}
\source{
\url{https://osf.io/8cd4r/}
}
\usage{
ML2.S2
}
\description{
A dataset containing results from the second slate of the ManyLabs2 project:
}
\details{
Klein, R. A., Vianello, M., Hasselman, F., Adams, B. G., Adams, R. B., Alper, S., ??? Friedman, M. (2016, March 15).
Many Labs 2: Investigating Variation in Replicability Across Sample and Setting. Retrieved from osf.io/8cd4r
}
\keyword{datasets}

