% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Alter.3}
\alias{varfun.Alter.3}
\title{varfun.Alter.3}
\usage{
varfun.Alter.3(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\url{http://fredhasselman.com/htmlHost/ManyLabs/ML2_PoPS_proposal.html#alter}
}
\details{
\strong{Analysis plan:} Similar to Alter et al. (2007), we will conduct an independent samples ttest to determine whether accuracy in solving moderately difficult syllogisms differ by font condition (fluent versus disfluent).

The original study focused on the moderately difficult questions, on the basis that participants' performance could vary enough to detect changes in processing depth.
Our primary analysis strategy will be sensitive to potential differences across samples in ability on syllogisms.
We will first determine which syllogisms were moderately difficult to participants by excluding any of the six items, within each sample, that were answered correctly by fewer than 25% of participants or more than 75\% of participants across conditions.
The remaining syllogisms will be the basis of computing mean syllogism performance for each participant. F
or a direct comparison with the original effect size, only English in-lab samples will be used for two reasons:
(1) we cannot adequately control for online participants "zooming in" on the page or otherwise making the font more readable, and
(2) a different font may be used in some translated versions because the original font (Myriad Web) may not support all languages.

All samples will be included in the investigation of cross-site variability in effect size.
}
\section{Variables}{


  Syllogisms to include for each sample
  INCLUSION PERCENTAGE BASED ON
   FLUENT / DISFLUENT SEPERATELY: 1 5 6
   BOTH: 1 5 6
   @examples
}

