% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Bauer.1}
\alias{varfun.Bauer.1}
\title{varfun.Bauer.1}
\usage{
varfun.Bauer.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\url{http://fredhasselman.com/htmlHost/ManyLabs/ML2_PoPS_proposal.html#bauer}
}
\details{
\strong{Analysis plan:} We will compare the mean trust levels between conditions with an independent samples t-test.
All participants with data will be included in analysis
Known differences from original. The original experiment included four additional dependent Variables:
(1) responsibility for the crisis,
(2) obligation to cut water usage,
(3) howmuch they viewed others as partners, and (4) how much others should use less water.

The central replication will be on the trust variable, while the other four dependent Variables: will be retained in the procedure but not analyzed for the focal replication.
}

