% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Gati.2}
\alias{varfun.Gati.2}
\title{varfun.Gati.2}
\usage{
varfun.Gati.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\strong{Analysis plan:} We will perform three analyses on the data.

For the primary analysis, we will analyze the data through a general linear mixed model with a random effect for the item pair
nested within subject, and a fixed factor 'order' representing the order of the pair (prominent first vs. prominent second).
Fitting this model will allow evaluation of both effects. If the intercept is significantly greater than 0, this would confirm the finding that at the participant level, if there is an effect for the factor 'order' the pairs where the prominent country appeared second will be rated as more similar than when the prominent country appeared first. We will convert the Beta provided by this intercept term into a Cohen???s d effect size.

Second, we will recreate the original analysis used to get a participant-level effect of making similarity judgments where either the more or less prominent country comes first. We will compute an asymmetry score for each subject, calculated as the average similarity for comparisons where the prominent country appears second minus the average for the comparisons where the prominent country appears first. Using a one-sample t-test, we will test this difference score against zero (original d=.48).

Third, using a matched-pairs t-test, we will compare the average score for each pair when it was prominent-first compared to prominent-second (original d = 1.31).

Because these latter two analyses do not account for the fact that the variance in ratings is
crossed between participants and pairs, they will be secondary and only used as a comparison for
the original analysis. All participants with data will be included in the analysis.

These analyses will be repeated for the differences conditions and reported as a separate study. Because of the random assignment to similarity or difference conditions, each site will have half as much data for its critical test as the other effects. This will likely increase the standard error of its estimates by comparison.
}
\references{
Tversky, A., \& Gati, I. (1978). Studies of similarity. \strong{Cognition and categorization}, 1, 79-98.
}

