% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Giessner.1}
\alias{varfun.Giessner.1}
\title{varfun.Giessner.1}
\usage{
varfun.Giessner.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
varfun.Giessner.1
}
\details{
\strong{Analysis plan:} Responses to the five dependent measures will be averaged and an independent samples
t-test will compare mean power rating between the 2 cm and 7 cm conditions.
All participants who complete at least one item of the dependent measure will be  included in the analysis.
}
\section{Variables}{

geis.1.1=long line condition;
geis.2.1=short line condition;
geis.dv_1=dominant;
geis.dv_2=strong;
geis.dv_3=self-confident;
geis.dv_4=control;
geis.dv_5=status;
For all dvs, higher numbers=higher power.
}

