% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Graham.1}
\alias{varfun.Graham.1}
\title{varfun.Graham.1}
\usage{
varfun.Graham.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\url{http://fredhasselman.com/htmlHost/ManyLabs/ML2_PoPS_proposal.html#graham}

The 9 items for ingroup, authority, and purity will be averaged to create a "binding" foundations moral relevance score.
The relationship between political ideology and the "binding" aggregates will be calculated using zero order correlations.

 The primary target of replication is the relationship of political ideology with the "binding" foundations.

 All participants who complete the corresponding measures will be included in analysis.
}

