% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Graham.2}
\alias{varfun.Graham.2}
\title{varfun.Graham.2}
\usage{
varfun.Graham.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\url{http://fredhasselman.com/htmlHost/ManyLabs/ML2_PoPS_proposal.html#graham}
}
\details{
The 6 items for harm and fairness will be averaged to create a "individualizing" foundations moral relevance score.
The relationship between political ideology and "individualizing" aggregates will be calculated using zero order correlations.

 The relationship of political ideology with the "individualizing" foundations is a secondary replication.

 All participants who complete the corresponding measures will be included in analysis.
}

