% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Gray.2}
\alias{varfun.Gray.2}
\title{varfun.Gray.2}
\usage{
varfun.Gray.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
varfun.Gray.2
}
\details{
\strong{Analysis plan:} We will compare the means on perceived responsibility between conditions with an independent samples t-test for the responsibility item.
The intentionality item will be analyzed the same way as a secondary analysis.
All participants with data will be included in analysis.
}
\section{Variables}{

Baby harms adult scenario; gray1.3=intentionality (adult); gray1.4=pain (baby).
Adult harms baby scenario; gray2.3=intentionality (baby); gray2.4=pain (adult)
}
\references{
Gray, K., & Wegner, D. M. (2009). Moral typecasting: divergent perceptions of moral agents and moral patients. Journal of Personality and Social Psychology, 96, 505.
}

