% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Hauser.1}
\alias{varfun.Hauser.1}
\title{varfun.Hauser.1}
\usage{
varfun.Hauser.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
varfun.Hauser.1
}
\details{
\strong{Analysis plan:} Subjects will be excluded from all analyses if they take fewer than four
seconds to read and respond to either of the target scenarios. For the key confirmatory test
comparing with the original effect size, we will include only participants that indicate having no
prior experience with the task. The original authors suggested the effect may be weaker in
participants with prior exposure. Prior exposure will be investigated as a moderator for the other
analyses. A two-way contingency table will be built with Scenario (Means vs. Side-effect) and
Response (Yes vs. No) as factors. The critical replication hypothesis will be given by a one tailed
chi square test and the effect size by an odds ratio.
}
\section{Variables}{

  haus1.1t = timing (side effect scenario);
  haus2.1t = timing (greater good scenario);
  haus1.2=previous experience (drop if 1 (yes));
  haus2.2=previous experience (drop if 1 (yes));
  haus1.1=morally permissible (side effect scenario; Yes=1);
  haus2.1=morally permissible (greater good scenario; Yes=1).
}

