% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Hauser.2}
\alias{varfun.Hauser.2}
\title{varfun.Hauser.2}
\usage{
varfun.Hauser.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
varfun.Hauser.2
}
\details{
\strong{Analysis Plan:} Participants will be excluded from all analyses if they take fewer than four seconds to read and respond to either of the target scenarios.
  For the key confirmatory test comparing with the original effect size, we will include only participants that indicate having no prior experience with the task.
  Prior exposure will be investigated as a moderator for the other analyses
}
\section{Variables}{

hauser3.1=morality judgment (greater good condition);
hauser4.1=morality judgment (foreseen side-effect condition; for both, yes=1, no=2.)
haus3.2 and haus4.2=previous experience (yes=1, no=2);
haus3.1t_3=timing (greater good);
haus4.1t_3=timing (side effect).
}
\references{
Hauser, M., Cushman, F., Young, L., Kang-Xing Jin, R., & Mikhail, J. (2007). A dissociation between moral judgments and justifications. Mind & Language, 22, 1-21.
}

