% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Hsee.1}
\alias{varfun.Hsee.1}
\title{varfun.Hsee.1}
\usage{
varfun.Hsee.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
varfun.Hsee.1
}
\details{
\strong{Analysis plan:} The two conditions will be compared with an independent samples t-test with rated generosity of gift giver as the dependent variable.
All participants with data will be included in the analysis.
}
\section{Variables}{

hsee1.1=generosity ($90 scarf condition);
hsee2.1=generosity ($110 coat condition);
for both, higher numbers=higher generosity
}
\references{
Hsee, C. K. (1998). Less is better: When low-value options are valued more highly than high-value options. Journal of Behavioral Decision Making, 11, 107-121.
}

