% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Huang.1}
\alias{varfun.Huang.1}
\title{Huang.1}
\usage{
varfun.Huang.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
A \code{list} object containing fields \strong{High}, \strong{Low} and \strong{N}
}
\description{
\url{http://fredhasselman.com/htmlHost/ManyLabs/ML2_PoPS_proposal.html#huang}
}
\details{
\strong{Analysis plan:} The coordinates of the click on the map will be recorded (X, Y) from the top-left of the image. The mean difference between the high and low-SES conditions for north/south location of click (Y) will be compared with an independent samples t-test. All participants who indicate an area within the boundaries of the map will be included in the analysis.
}
\section{Variables}{


huan1.1_Y1: Y position of the mouse (High SES condition).
huan2.1_Y1: Y position of the mouse (Low SES).
huan1.1_R0 and huan2.1_R0 indicate for each condition whether a click was inside the map (1) or outside (0).

For each condition a participant must have clicked inside the map (=1) to be included in the analysis.
}

