% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Inbar.1}
\alias{varfun.Inbar.1}
\title{varfun.Inbar.1}
\usage{
varfun.Inbar.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\url{http://fredhasselman.com/htmlHost/ManyLabs/ML2_PoPS_proposal.html#inbar}
}
\details{
\strong{Analysis plan:} The five items of the contamination subscale of the Disgust Sensitivity Scale-Revised will be averaged to create a single index of disgust sensitivity.
For the primary analysis, we will compute a correlation between disgust sensitivity and assessments of the director's intentionality in both the gay kissing and kissing conditions,
and then compare the correlations with an r-to-z transformation.
The other two outcome measures will be examined as secondary analyses following the same analysis strategy.
All participants with relevant data will be included in analysis.
}
\section{Variables}{

disg1.11,disg1.12,disg2.10,disg2.12,disg2.13;

responses on the DS-R are scored as follows: True 1, False 0; Not disgusting 0, Slightly disgusting 0.5, Very disgusting 1
}

