% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Kay.1}
\alias{varfun.Kay.1}
\title{varfun.Kay.1}
\usage{
varfun.Kay.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\url{http://fredhasselman.com/htmlHost/ManyLabs/ML2_PoPS_proposal.html#kay}
}
\details{
\strong{Analysis plan:} Following Kay et al. (2014), we will create an index of willingness to engage in goal pursuit for each participant by
(1) regressing the mean of the two goal pursuit items on the centered mean of the goal subjective value item,
(2) calculating the unstandardized residual for each participant, and
(3) add to those the mean value for the self-regulation items measuring willingness to engage in goal pursuit.

Then, the two conditions will be compared using an independent samples t-test.
}

