% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Knobe.2}
\alias{varfun.Knobe.2}
\title{varfun.Knobe.2}
\usage{
varfun.Knobe.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
varfun.Knobe.2
}
\details{
\strong{Analysis plan:} Ratings of intentionality in the harm and help conditions will be
compared using an independent-samples t-test. This is the focal test for the direct replication.
Blame and praise ratings will also be collected but are secondary analyses. All participants with
data will be included in analysis.
}
\section{Variables}{

  knob1.4=intentionality (praise condition);
  knob2.4=intentionality (blame condition);
  for both, higher numbers=higher intentionality
}
\references{
Knobe, J. (2003). Intentional action and side effects in ordinary language. Analysis, 63, 190-193.
}

