% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Miyamoto.1}
\alias{varfun.Miyamoto.1}
\title{varfun.Miyamoto.1}
\usage{
varfun.Miyamoto.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\url{http://fredhasselman.com/htmlHost/ManyLabs/ML2_PoPS_proposal.html#Miyamoto}
}
\details{
\strong{Analysis plan:} An ANCOVA will compare the mean estimates of the author's true attitude across the two conditions, covarying for perceived constraint.
}
\section{Variables}{

miya1.5=true attitude (pro-death condition; higher values=higher support for death penalty);
miya1.7=perceived constraint (pro-death condition; higher values=higher freedom);

miya2.5=true attitude (against death penalty condition; higher values=higher support for death penalty);
miya2.7=perceived constraint (against death condition; higher values= higher freedom).
}

