% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Norenzayan.1}
\alias{varfun.Norenzayan.1}
\title{varfun.Norenzayan.1}
\usage{
varfun.Norenzayan.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{

}
\description{
varfun.Norenzayan.1
}
\details{
\strong{Analysis plan:} We will compute for each subject the percentage of rule-based responses and test whether the mean of the two experimental groups (belong vs similar) on this DV is equal with a t-test for independent samples.
 The effect size will be given by a standardized mean difference. All participants with data will be included in analysis.

For additional analysis, a few items about cultural origins of participants and their parents are present in the individual differences assessment.
  These could be particularly useful for follow-up moderator analysis.
}
\section{Variables}{

 nore1.1 TO nore1.20 provide choices ("belong to" condition)

 nore2.1 to nore2.20 provide choices ("similar to" condition)
}
\references{
Norenzayan, A., Smith, E. E., Kim, B. J., & Nisbett, R. E. (2002). Cultural preferences for formal versus intuitive reasoning. Cognitive Science, 26, 653-684.
}

