% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Norenzayan.2}
\alias{varfun.Norenzayan.2}
\title{varfun.Norenzayan.2}
\usage{
varfun.Norenzayan.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
varfun.Norenzayan.2
}
\details{
\strong{Analysis plan:} We will compute for each subject the percentage of rule-based responses and test whether the mean of the two experimental groups (belong vs similar) on this DV is equal with a t-test for independent samples.
The effect size will be given by a standardized mean difference.
All participants with data will be included in analysis.
}
\note{
This analysis tests moderating effect of presenting Gati first or after.
}
\references{
Norenzayan, A., Smith, E. E., Kim, B. J., & Nisbett, R. E. (2002). Cultural preferences for formal versus intuitive reasoning. Cognitive Science, 26, 653-684.
}

