% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Risen.2}
\alias{varfun.Risen.2}
\title{varfun.Risen.2}
\usage{
varfun.Risen.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
varfun.Risen.2
}
\details{
\strong{Analysis plan:} The two groups will be compared with an independent samples t-test. All participants that answer the dependent measure will be included in analysis.
 The primary confirmatory test for comparing the original and replication effect size will be based on only the samples using undergraduate students.
 We will examine gender as a possible moderator of the effect in a supplemental, exploratory analysis.
 use variable "sex" 1 = male, 2 = female. (3 = other and 4 = prefer not to answer, which I think we probably do not use for this analysis)
}
\section{Variables}{

rise1.3=likelihood that the professor will call on you (unprepared condition);
rise2.3=likelihood that the professor will call on you (prepared condition);

for both, higher numbers=higher likelihood   All participants that answer the dependent measure will be included in analysis.
The primary confirmatory test for comparing the original and replication effect size will be based on only the samples using undergraduate students.
}
\references{
Risen, J. L., & Gilovich, T. (2008). Why people are reluctant to tempt fate. Journal of Personality and Social Psychology, 95, 293.
}

