% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Ross.1}
\alias{varfun.Ross.1}
\title{varfun.Ross.1}
\usage{
varfun.Ross.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
varfun.Ross.1
}
\details{
\strong{Analysis plan:} An independent samples t-test will be conducted with participants' choice
(sign release/refuse to sign) as the IV and participant estimate of % of peers who would sign the
release as the DV.

Note that participants self-select whether to sign or refuse to sign the release, so it is not random assignment to levels of the independent variable.
Participants will be included in the analysis if they respond to all three questions and their estimate for the DV (e.g., 'what percent of your peers would sign the release') falls between 0-100
}
\section{Variables}{

ross.s1.1 = percentage of peers;
ross.s1.2 = you; values: 1=sign; 2=refuse
}

