% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Rottenstreich.1}
\alias{varfun.Rottenstreich.1}
\title{varfun.Rottenstreich.1}
\usage{
varfun.Rottenstreich.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\url{http://fredhasselman.com/htmlHost/ManyLabs/ML2_PoPS_proposal.html#rottenstreich}
}
\details{
\strong{Analysis plan:} A two-way contingency table will be built with certainty condition (lowprobability
vs. certain) and choice (monetary reward vs. meeting favorite movie star) as factors.
The critical replication hypothesis will be given by a X2 test and the effect size by an odds ratio.
All participants with valid data on the response will be included in the analysis.
}

