% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Shafir.1}
\alias{varfun.Shafir.1}
\title{varfun.Shafir.1}
\usage{
varfun.Shafir.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
varfun.Shafir.1
}
\details{
\strong{Analysis plan:} The proportion of participants awarding or denying custody for parent B
will be summed from both groups and tested against 100% with a Z test. The effect size will be
computed estimating a logistic regression model on the 2X2 table and then taking the
exponentiation of the unstandardized beta parameter of the main effect of parent B, which can be
interpreted as an odds ratio. All participants with data will be included in analysis.
}
\section{Variables}{

shaf1.1=choice (award condition; Parent A=1, Parent B=2);
shaf2.1=choice (deny condition; Parent A=1, Parent B=2)
}
\references{
Shafir, E. (1993). Choosing versus rejecting: Why some options are both better and worse than others. Memory & Cognition, 21, 546-556.
}

