% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.vanLange.1}
\alias{varfun.vanLange.1}
\title{van.Lange.1}
\usage{
varfun.vanLange.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
A \code{list} object containing fields \strong{SVO}, \strong{Siblings} and \strong{N}
}
\description{
\url{http://fredhasselman.com/htmlHost/ManyLabs/ML2_PoPS_proposal.html#vanlange}
}
\details{
\strong{Analysis plan:} The current replication focuses only on the observed direct positive correlation between greater prosocial orientation and number of siblings.

  Participants must respond to all 6 items of the SVO slider and have valid data for the two questions asking about older and younger siblings, to be included in the analysis. Total number of siblings will be obtained by adding the number of younger and older siblings.

  SVO slider scores will be scored in according to the procedure recommended by Murphy et al. (2011). The resulting SVO slide score will be correlated with the total number of siblings for the critical test.
}
\section{Variables}{

   van.p.1.2_1 TO van.p.1.2_6 are the items of SVO measure.

  murphy et al. (2011) scoring: SVO degress=arctan [(mean Alloc other - 50)/(mean Allocation self - 50)].

  See SVO codes (this doc) for the list of paired amounts
  van.p2.1_1_TEXT= # of older siblings;
  van.p2.1_2_TEXT= # of younger siblings.
}

