% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Zaval.1}
\alias{varfun.Zaval.1}
\title{varfun.Zaval.1}
\usage{
varfun.Zaval.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\strong{Analysis plan:} Mean differences in belief and concern about global warming between heat and cold-priming conditions will be evaluated with an independent samples t-test. The scrambled sentence task could introduce unanticipated variation across translations. As such, the direct replication test will use only English language sites, and - like all other effects - all samples and settings with data will be included in analyses examining heterogeneity to see if factors, like translation, have an impact on effect estimates.

the direct replication test will use only English language sites,
and - like all other effects - all samples and settings with data will be included in analyses examining heterogeneity to see if factors, like translation,
have an impact on effect estimates
}
\section{Variables}{

zav1.1 TO zav1.13 provide COLD primes.
zav2.1 TO zav2.13 provide HEAT primes.
zav.dv.2=belief;
zav.dv.3=concern;
higher numbers=higher belief/concern.
}
\references{
Zaval, L., Keenan, E. A., Johnson, E. J., & Weber, E. U. (2014). How warm days increase belief in global warming. \strong{Nature Climate Change}, 4, 143-147.
}

