% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Zhong.1}
\alias{varfun.Zhong.1}
\title{varfun.Zhong.1}
\usage{
varfun.Zhong.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
varfun.Zhong.1
}
\details{
\strong{Analysis plan:} The key factor of interest is whether condition affects ratings of the cleaning products, so ratings of the five cleaning products will be averaged and compared between the two conditions (ethical or unethical) with an independent samples t-test.

A second comparison is whether there is a condition difference between ratings of the other products. The theoretical expectation is that this difference will be weak or near zero. This will be examined as a secondary analysis as a 2 (ethical-unethical) x 2 (cleaning-other products) mixed model ANOVA, and a follow-up independent samples t-test comparing ratings of other products between the ethical and unethical conditions.3

Participants who copy less than half the target article will be excluded from analysis.
}
\section{Variables}{

zhon1.1= unethical condition;
zhon2.1=ethical condition;

zhon.dv.1_1 TO zhon.dv.1_10=desirability of products (both conditions);
higher numbers=higher desirability;

Products list:

Clean:
Dove shower soap (zhon.dv.1_2),
Crest toothpaste (zhon.dv.1_3),
Windex glass cleaner (zhon.dv.1_7),
Lysol countertop disinfectant (zhon.dv.1_8),
Tide laundry detergent (zhon.dv.1_10)

Not-clean:
Post-it notes (zhon.dv.1_1),
Nantucket Nectars juice (zhon.dv.1_4),
Energizer batteries (zhon.dv.1_5),
Sony cd cases (zhon.dv.1_6),
Snickers candy bar (zhon.dv.1_9),
}
\references{
Zhong, C. B., & Liljenquist, K. (2006). Washing away your sins: Threatened morality and physical cleansing. Science, 313, 1451???1452.
}

